pyproject.toml
[tool.poetry]
name = "is3-broker-rl"
version = "0.1.0"
description = "A Python starter template with Poetry and tox."
authors = ["Philipp Page <github@philipp-page.de>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
ray = { extras = ["rllib", "serve", "tune"], version = "1.12.1" }
protobuf = "~3.19.0"
tensorflow = { version = "^2.9.0", markers = "sys_platform == 'win32' or sys_platform == 'linux' or (sys_platform == 'darwin' and platform_machine != 'arm64')" }
tensorflow-macos = { version = "^2.9.0", markers = "sys_platform == 'darwin' and platform_machine == 'arm64'" }
python-dotenv = "^0.20.0"
gym = "^0.21.0"
matplotlib = "^3.5.2"
seaborn = "^0.11.2"
numpy = "^1.22.4"
pandas = "^1.4.2"

[tool.poetry.dev-dependencies]
pytest = "^7.0.0"
tox = "^3.24.4"
isort = "^5.10.1"
black = "^22.1.0"
flake8 = "^4.0.1"
mypy = "^0.961"
pytest-cov = "^3.0.0"
tensorboard = "^2.9.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"

[tool.mypy]
files = "is3_broker_rl"
ignore_missing_imports = true
plugins = "numpy.typing.mypy_plugin"
